server/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
django-redis==5.4.0
django-celery-results==2.5.1
django-filter==23.5
psycopg2-binary==2.9.11
redis==5.0.1
openai==2.8.1
cloudinary==1.36.0
python-dotenv==1.0.0
gunicorn==21.2.0
dj-database-url==2.1.0
Pillow>=10.3.0
Jinja2==3.1.2
celery==5.5.3
setuptools
wheel